# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reportist API",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "ReportMeta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "body-composition-dynamics"
          },
          "name": {
            "type": "string",
            "example": "Body composition dynamics"
          },
          "description": {
            "type": "string"
          },
          "supportedFormats": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "xlsx"
              ]
            }
          },
          "paramsSchema": {
            "type": "object",
            "description": "JSON Schema (draft 2020-12) of report parameters"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "supportedFormats"
        ]
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "gymId": {
            "type": "string",
            "format": "uuid"
          },
          "gymName": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "gymId",
          "gymName",
          "city"
        ]
      },
      "CreateRunResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ]
          }
        },
        "required": [
          "id",
          "status"
        ]
      },
      "CreateRunBody": {
        "type": "object",
        "properties": {
          "reportId": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "enum": [
              "xlsx"
            ]
          },
          "params": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "reportId",
          "format",
          "params"
        ]
      },
      "Run": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reportId": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "enum": [
              "xlsx"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ]
          },
          "params": {
            "type": "object",
            "additionalProperties": {}
          },
          "createdAt": {
            "type": "string"
          },
          "resultKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "reportId",
          "format",
          "status",
          "params",
          "createdAt",
          "resultKey",
          "errorMessage"
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/api/reports": {
      "get": {
        "operationId": "listReports",
        "tags": [
          "reports"
        ],
        "summary": "List available reports",
        "responses": {
          "200": {
            "description": "Registered report definitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportMeta"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/clients": {
      "get": {
        "operationId": "listClients",
        "tags": [
          "clients"
        ],
        "summary": "List fitness clients available for reports",
        "responses": {
          "200": {
            "description": "Clients with their gym context",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "post": {
        "operationId": "createRun",
        "tags": [
          "runs"
        ],
        "summary": "Enqueue a new report run",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Run has been queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid report id, format or params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listRuns",
        "tags": [
          "runs"
        ],
        "summary": "Current user's run history",
        "responses": {
          "200": {
            "description": "Runs ordered by creation time desc",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "operationId": "getRun",
        "tags": [
          "runs"
        ],
        "summary": "Single run for polling",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Single run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "webhooks": {}
}